Joint Random Bit Sharing.

+++ common.cpp
@@ -1,4 +1,6 @@
 
-int mod(int x, int m) {
+#include "common.hpp"
+
+int mod( int64_t x, int m) {
     return (x%m + m)%m;
 }
