+++ peer/peer.hpp
@@ -14,7 +14,7 @@
 #include <tbb/concurrent_unordered_map.h>
 #include <tbb/mutex.h>
 
-typedef int plaintext_t;
+typedef  int64_t plaintext_t;
 typedef string symbol_t;
 
 class RPCServer;
@@ -26,7 +26,7 @@ public:
 
   template<class PeerSeq>
   void distribute_secret(
-      const symbol_t key, const int value,
+      const symbol_t key, const int64_t value,
       PeerSeq comp_peers);
 
   template<class Values, class PeerSeq>
@@ -34,11 +34,11 @@ public:
       const symbol_t key, const Values values,
       PeerSeq comp_peers);
 
-  void publish(std::string key, int value);
+  void publish(std::string key,  int64_t value);
   void print_values();
 
-  typedef tbb::concurrent_unordered_map<int, int> inter_map_t;
-  typedef tbb::concurrent_unordered_map <symbol_t, int> value_map_t;
+  typedef tbb::concurrent_unordered_map<int, int64_t> inter_map_t;
+  typedef tbb::concurrent_unordered_map <symbol_t, int64_t> value_map_t;
 
   io_service io_service_;
   RPCServer* server_;
